+++ database/migrations/2023_06_20_181003_create_transactions_table.php
@@ -17,7 +17,7 @@ class CreateTransactionsTable extends Migration
         Schema::create('transactions', function (Blueprint $table) {
             $table->id();
             $table->foreignIdFor(User::class)->nullable();
-            $table->integer('total');
+            $table->double('total');
             $table->timestamps();
         });
     }
